==> chap_11/03_HTRF.ipynb <==
# How to Train & Tune a Random Forest

### Loading Libraries

# OS
import os, sys

# Numerical Computing
import numpy as np
from numpy.random import choice, normal

# Data Manipulation
import pandas as pd

# Data Visualization
import seaborn as sns
import matplotlib.pyplot as plt

# SciPy
from scipy.stats import spearmanr

# Warnings
import warnings

# Path
from pathlib import Path

# JobLib
import joblib

# Scikit-Learn
from sklearn.metrics import make_scorer
from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import GridSearchCV, cross_val_score
from sklearn.ensemble import BaggingRegressor, RandomForestRegressor, RandomForestClassifier

# Utils- Outside Library
from utils import MultipleTimeSeriesCV

sys.path.insert(1, os.path.join(sys.path[0], '..'))

sns.set_style('white')

np.random.seed(seed=42)

warnings.filterwarnings('ignore')

results_path = Path('results', 'random_forest')

if not results_path.exists():
    results_path.mkdir(parents=True)

### Getting Data

with pd.HDFStore('data.h5') as store:
    data =store['us/equities/monthly']

data.info()

y = data.target
y_binary = (y > 0).astype(int)

X = pd.get_dummies(data.drop('target', axis=1))

### Random Forests

#### Cross-Validation Parameters

n_splits = 10
train_period_length = 60
test_period_length = 6
lookahead = 1

cv = MultipleTimeSeriesCV(n_splits=n_splits,
                          train_period_length=train_period_length,
                          test_period_length=test_period_length,
                          lookahead=lookahead)

#### Classifier

rf_clf = RandomForestClassifier(n_estimators=100,       
                                criterion='gini', 
                                max_depth=None, 
                                min_samples_split=2, 
                                min_samples_leaf=1, 
                                min_weight_fraction_leaf=0.0, 
                                max_features='auto',
                                max_leaf_nodes=None, 
                                min_impurity_decrease=0.0, 
                                min_impurity_split=None, 
                                bootstrap=True, 
                                oob_score=True, 
                                n_jobs=-1,
                                random_state=42, 
                                verbose=1)

#### Cross-Validation with Default Settings

cv_score = cross_val_score(estimator=rf_clf,
                           X=X,
                           y=y_binary,
                           scoring='roc_auc',
                           cv=cv,
                           n_jobs=-1,
                           verbose=1)

np.mean(cv_score)

#### Regression RF

def rank_correl(y, y_pred):
    return spearmanr(y, y_pred)[0]

ic = make_scorer(rank_correl)

rf_reg = RandomForestRegressor(n_estimators=100, 
                                max_depth=None, 
                                min_samples_split=2, 
                                min_samples_leaf=1, 
                                min_weight_fraction_leaf=0.0, 
                                max_features='auto', 
                                max_leaf_nodes=None, 
                                min_impurity_decrease=0.0, 
                                min_impurity_split=None, 
                                bootstrap=True, 
                                oob_score=False, 
                                n_jobs=-1, 
                                random_state=None, 
                                verbose=0, 
                                warm_start=False)

cv_score = cross_val_score(estimator=rf_reg,
                           X=X,
                           y=y,
                           scoring=ic,
                           cv=cv,
                           n_jobs=-1,
                           verbose=1)

np.mean(cv_score)

### Parameter Tuning

#### Defining Parameter Grid

param_grid = {'n_estimators': [50, 100, 250],
              'max_depth': [5, 15, None],
              'min_samples_leaf': [5, 25, 100]}

gridsearch_clf = GridSearchCV(estimator=rf_clf,
                              param_grid=param_grid,
                              scoring='roc_auc',
                              n_jobs=-1,
                              cv=cv,
                              refit=True,
                              return_train_score=True,
                              verbose=1)

#### Fitting Classifier

gridsearch_clf.fit(X=X, y=y_binary)

#### Persisting Result

joblib.dump(gridsearch_clf, results_path / 'gridsearch_clf.joblib')

gridsearch_clf = joblib.load(results_path / 'gridsearch_clf.joblib')

gridsearch_clf.best_params_

gridsearch_clf.best_score_

#### Feature Importance

fig, ax = plt.subplots(figsize=(12,5))

(pd.Series(gridsearch_clf.best_estimator_.feature_importances_, 
           index=X.columns)
 .sort_values(ascending=False)
 .iloc[:20]
 .sort_values()
 .plot.barh(ax=ax, title='RF Feature Importance'))
sns.despine()
fig.tight_layout();
plt.show()

#### Fitting Regressor

gridsearch_reg = GridSearchCV(estimator=rf_reg,
                      param_grid=param_grid,
                      scoring=ic,
                      n_jobs=-1,
                      cv=cv,
                      refit=True,
                      return_train_score=True,
                      verbose=1)

gs_reg = gridsearch_reg

gridsearch_reg.fit(X=X, y=y)

joblib.dump(gridsearch_reg, results_path / 'rf_reg_gridsearch.joblib')

gridsearch_reg = joblib.load(results_path / 'rf_reg_gridsearch.joblib')

gridsearch_reg.best_params_

f'{gridsearch_reg.best_score_*100:.2f}'

### Comparing Results

#### Best Parameters

pd.DataFrame({'Regression': pd.Series(gridsearch_reg.best_params_),
              'Classification': pd.Series(gridsearch_clf.best_params_)})

fi_clf = gridsearch_clf.best_estimator_.feature_importances_

fi_reg = gridsearch_reg.best_estimator_.feature_importances_

idx = [c.replace('_', ' ').upper() for c in X.columns]

fig, axes = plt.subplots(figsize=(14, 4), ncols=2)

(pd.Series(fi_clf, index=idx)
 .sort_values(ascending=False)
 .iloc[:15]
 .sort_values()
 .plot.barh(ax=axes[1], title='Classifier'))
(pd.Series(fi_reg, index=idx)
 .sort_values(ascending=False)
 .iloc[:15]
 .sort_values()
 .plot.barh(ax=axes[0], title='Regression'))
sns.despine()
fig.tight_layout()
plt.show()
