==> chap_11/06_ASQT.ipynb <==
# Testing The Signal Quality with Alphalens

### Loading Libraries

# Warnings
import warnings

# Numerical Computing
import numpy as np

# Data Manipulation
import pandas as pd

# Data visualization
import seaborn as sns
import matplotlib.pyplot as plt

# Technical Analysis
import talib

# SciPy
from scipy.stats import spearmanr

# Path
from pathlib import Path
from itertools import product

# OS & Time
import sys, os
from time import time
from io import StringIO

# Notebook Optimizer
from tqdm import tqdm

# StatsModel
import statsmodels.api as sm

# Scikit-Learn
from sklearn.linear_model import LinearRegression

# LightBoosting Gradient
import lightgbm as lgb

# AlphaLens
from alphalens.tears import (create_summary_tear_sheet,
                             create_full_tear_sheet)

from alphalens.utils import get_clean_factor_and_forward_returns

np.random.seed(42)

sns.set_style('white')

warnings.filterwarnings('ignore')

idx = pd.IndexSlice

results_path = Path('results', 'return_predictions')

if not results_path.exists():
    results_path.mkdir(parents=True)

### Evaluating The Cross-Validation Results

lookahead = 1

cv_store = Path(results_path / 'parameter_tuning.h5')

### Getting AlphaLens Input

DATA_DIR = Path('..', 'data')

def get_trade_prices(tickers):
    store = DATA_DIR / 'assets.h5'
    prices = pd.read_hdf(store, 'stooq/jp/tse/stocks/prices')
    return (prices.loc[idx[tickers, '2014': '2019'], 'open']
            .unstack('ticker')
            .sort_index()
            .shift(-1)
            .dropna()
            .tz_localize('UTC'))

best_predictions = pd.read_hdf(results_path / 'predictions.h5', f'test/{lookahead:02}')
best_predictions.info()

test_tickers = best_predictions.index.get_level_values('ticker').unique()

trade_prices = get_trade_prices(test_tickers)
trade_prices.info()

factor = (best_predictions
          .iloc[:, :3]
          .mean(1)
          .tz_localize('UTC', level='date')
          .swaplevel()
          .dropna()
          .reset_index()
          .drop_duplicates()
          .set_index(['date', 'ticker']))

factor_data = get_clean_factor_and_forward_returns(factor=factor,
                                                   prices=trade_prices,
                                                   quantiles=5,
                                                   periods=(1, 5, 10, 21))
factor_data.sort_index().info()

### Summary Tearsheet

create_summary_tear_sheet(factor_data)

### Evaluating The Out-of-Sample Predictions

#### Preparing Factor Data

t = 1

predictions = pd.read_hdf(results_path / 'predictions.h5',
                          f'test/{t:02}').drop('y_test', axis=1)

predictions.info()

factor = (predictions.iloc[:, :10]
                   .mean(1)
                   .sort_index().tz_localize('UTC', level='date').swaplevel().dropna())

factor.head()

### Next Available Trade Prices Selection

tickers = factor.index.get_level_values('ticker').unique()
trade_prices = get_trade_prices(tickers)

trade_prices.info()

### Getting AlphaLens Inputs

factor_data = get_clean_factor_and_forward_returns(factor=factor,
                                                   prices=trade_prices,
                                                   quantiles=5,
                                                   periods=(1, 5, 10, 21))

factor_data.sort_index().info()

### Summary Tearsheet

create_summary_tear_sheet(factor_data)
